Combine 2 functions used for cleaning data into 1

+++ src/import_clean_data/clean_data.py
@@ -1,7 +1,6 @@
 import pandas as pd
 def clean_data(data: pd.DataFrame) -> pd.DataFrame:
-    data['Age']=data['Age'].fillna(data['Age'].mean().round())
-    data["Age"]=data["Age"].astype(int)
+    data['Age'] = data['Age'].fillna(data['Age'].mean().round()).astype(int)
     data['Sex'] = data['Sex'].map({'male':0,'female':1})
     data = data.dropna(subset=['Embarked'])
     data.loc[:, 'Embarked']= data['Embarked'].map({'S': 1, 'C': 2, 'Q': 3})
@@ -16,8 +15,7 @@ def clean_data(data: pd.DataFrame) -> pd.DataFrame:
     return data
 
 def clean_test_data(data: pd.DataFrame) -> pd.DataFrame:
-    data['Age'] = data['Age'].fillna(data['Age'].mean().round())
-    data["Age"] = data["Age"].astype(int)
+    data['Age'] = data['Age'].fillna(data['Age'].mean().round()).astype(int)
     data['Sex'] = data['Sex'].map({'male': 0, 'female': 1})
     data = data.dropna(subset=['Fare'])
     data.loc[:, 'Embarked'] = data['Embarked'].map({'S': 1, 'C': 2, 'Q': 3})
